Extract UserPreferences Component to separate file
